feat: Footer 폴더 밖으로 빼내기 & LinkIcon 사용
