--- app/_components/common/BackLink/index.tsx
@@ -1,37 +1,0 @@
-"use client";
-
-import Image from "next/image";
-import React from "react";
-
-interface Props {
-	children: string;
-}
-
-export default function BackLink({ children }: Props) {
-	const copyTextToClipboard = async (text: string) => {
-		try {
-			await navigator.clipboard.writeText(text); // 클립보드에 복사
-
-			// 상태 표시 코드
-			//
-		} catch (error) {
-			console.error("클립보드 복사 실패");
-		}
-	};
-
-	const onClickCopy = () => {
-		copyTextToClipboard(children);
-	};
-
-	return (
-		<button className="flex items-center gap-[5px] rounded-[10px] bg-[#EEF9F6] p-[5px_10px] text-md text-[#4CBFA4]" onClick={onClickCopy}>
-			<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-				<path
-					d="M11.7017 18.7169C10.8138 19.6048 9.74414 20.0488 8.49265 20.0489C7.24115 20.049 6.17152 19.6051 5.28376 18.7174C4.396 17.8296 3.95203 16.7598 3.95185 15.5081C3.95168 14.2563 4.39552 13.1865 5.28339 12.2986L7.4183 10.1637C7.56856 10.0135 7.74536 9.93837 7.94868 9.93843C8.15201 9.93846 8.32884 10.0136 8.47917 10.164C8.62949 10.3143 8.70459 10.491 8.70448 10.6942C8.70437 10.8974 8.62919 11.0741 8.47893 11.2244L6.3432 13.3601C5.74995 13.9533 5.45332 14.6692 5.45332 15.5078C5.45332 16.3464 5.75022 17.0625 6.34401 17.6563C6.93781 18.2501 7.65398 18.547 8.49254 18.547C9.3311 18.547 10.047 18.2504 10.6403 17.6571L12.776 15.5214C12.9262 15.3711 13.103 15.296 13.3063 15.2961C13.5097 15.2961 13.6865 15.3713 13.8368 15.5216C13.9872 15.672 14.0623 15.8487 14.0622 16.0519C14.062 16.2551 13.9869 16.4318 13.8366 16.582L11.7017 18.7169ZM10.4099 14.6511C10.2596 14.8014 10.0828 14.8765 9.87949 14.8764C9.67615 14.8764 9.49932 14.8012 9.349 14.6509C9.19867 14.5005 9.12357 14.3238 9.12369 14.1206C9.12379 13.9174 9.19897 13.7407 9.34923 13.5905L13.5919 9.34783C13.7421 9.19757 13.9189 9.12246 14.1223 9.1225C14.3256 9.12255 14.5024 9.19774 14.6527 9.34806C14.8031 9.49839 14.8782 9.67514 14.8781 9.8783C14.8779 10.0815 14.8028 10.2582 14.6525 10.4085L10.4099 14.6511ZM16.5834 13.8352C16.4332 13.9855 16.2564 14.0606 16.053 14.0605C15.8497 14.0605 15.6729 13.9853 15.5226 13.835C15.3722 13.6846 15.2971 13.5079 15.2973 13.3047C15.2974 13.1016 15.3725 12.9248 15.5228 12.7746L17.6585 10.6389C18.2518 10.0456 18.5484 9.3297 18.5484 8.49114C18.5484 7.65258 18.2515 6.9364 17.6577 6.34261C17.0639 5.74882 16.3478 5.45192 15.5092 5.45192C14.6706 5.45192 13.9547 5.74855 13.3615 6.3418L11.2258 8.47753C11.0755 8.62779 10.8987 8.70289 10.6954 8.70285C10.492 8.7028 10.3152 8.62762 10.1649 8.4773C10.0146 8.32697 9.93946 8.15021 9.93958 7.94704C9.93969 7.74387 10.0149 7.56716 10.1651 7.4169L12.3 5.28199C13.1879 4.39412 14.2576 3.95015 15.5091 3.95006C16.7606 3.94998 17.8302 4.39382 18.718 5.28158C19.6057 6.16934 20.0497 7.2391 20.0499 8.49086C20.0501 9.74261 19.6062 10.8124 18.7183 11.7003L16.5834 13.8352Z"
-					fill="#4CBFA4"
-				/>
-			</svg>
-			<span>{children}</span>
-		</button>
-	);
-}
